Get rid of connectorUtils and move to the respective router, fix imports
